+++ fltk/src/prelude.rs
@@ -88,8 +88,8 @@ pub unsafe trait WidgetExt {
     fn set_pos(&mut self, x: i32, y: i32);
     /// Set to dimensions width and height
     fn set_size(&mut self, width: i32, height: i32);
-    /// Sets the widget's label
-    /// labels support special symbols preceded by an `@` [sign](https://www.fltk.org/doc-1.3/symbols.png)
+    /// Sets the widget's label.
+    /// labels support special symbols preceded by an `@` [sign](https://www.fltk.org/doc-1.3/symbols.png).
     /// and for the [associated formatting](https://www.fltk.org/doc-1.3/common.html).
     fn set_label(&mut self, title: &str);
     /// Redraws a widget, necessary for resizing and changing positions
@@ -106,6 +106,10 @@ pub unsafe trait WidgetExt {
     fn width(&self) -> i32;
     /// Returns the height of the widget
     fn height(&self) -> i32;
+    /// Returns the width of the widget
+    fn w(&self) -> i32;
+    /// Returns the height of the widget
+    fn h(&self) -> i32;
     /// Returns the label of the widget
     fn label(&self) -> String;
     /// Measures the label's width and height
@@ -318,7 +322,7 @@ pub unsafe trait WidgetBase: WidgetExt {
     /// * `width` - The width of the widget
     /// * `heigth` - The height of the widget
     /// * `title` - The title or label of the widget
-    /// labels support special symbols preceded by an `@` [sign](https://www.fltk.org/doc-1.3/symbols.png)
+    /// labels support special symbols preceded by an `@` [sign](https://www.fltk.org/doc-1.3/symbols.png).
     /// and for the [associated formatting](https://www.fltk.org/doc-1.3/common.html).
     fn new(x: i32, y: i32, width: i32, height: i32, title: &str) -> Self;
     /// Deletes widgets and their children.
@@ -843,8 +847,6 @@ pub unsafe trait BrowserExt: WidgetExt {
     fn clear(&mut self);
     /// Returns the number of items
     fn size(&self) -> u32;
-    /// Set the number of items
-    fn set_size(&mut self, w: i32, h: i32);
     /// Select an item at the specified line
     /// Lines start at 1
     fn select(&mut self, line: u32);